Type warnings fixes

+++ GLUtil.h
@@ -308,7 +308,7 @@ public:
 		LogInfo("Compiling GLSL shader %s", szPath);
 		std::ifstream ifVertexShader(szPath, std::ios::binary);
 		ifVertexShader.seekg(0, std::ios::end);
-		nBytes = ifVertexShader.tellg();
+		nBytes = (int)ifVertexShader.tellg();
 		ifVertexShader.seekg(0, std::ios::beg);
 		psz = new char[nBytes+1];
 		ifVertexShader.read(psz, nBytes);
@@ -330,7 +330,7 @@ public:
 		LogInfo("Compiling GLSL shader %s", szPath);
 		std::ifstream ifFragmentShader(szPath, std::ios::binary);
 		ifFragmentShader.seekg(0, std::ios::end);
-		nBytes = ifFragmentShader.tellg();
+		nBytes = (int)ifFragmentShader.tellg();
 		ifFragmentShader.seekg(0, std::ios::beg);
 		psz = new char[nBytes];
 		ifFragmentShader.read(psz, nBytes);
